Improve failed fetch info
